duplicate auto upgrades server mock
