+++ integration/proxy/automaticupgrades_test.go
@@ -24,6 +24,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"net/http/httptest"
 	"net/url"
 	"path/filepath"
 	"testing"
@@ -35,7 +36,6 @@ import (
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/integration/helpers"
 	"github.com/gravitational/teleport/lib/automaticupgrades"
-	"github.com/gravitational/teleport/lib/automaticupgrades/basichttp"
 	"github.com/gravitational/teleport/lib/automaticupgrades/constants"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
 	"github.com/gravitational/teleport/lib/utils"
@@ -75,6 +75,39 @@ func createProxyWithChannels(t *testing.T, channels automaticupgrades.Channels) 
 	return cfg.Listeners.Web
 }
 
+// ServerMock is a HTTP server whose response can be controlled from the tests.
+// This is used to mock external dependencies like s3 buckets or a remote HTTP server.
+type ServerMock struct {
+	Srv *httptest.Server
+
+	t        *testing.T
+	code     int
+	response string
+	path     string
+}
+
+// SetResponse sets the ServerMock's response.
+func (m *ServerMock) SetResponse(t *testing.T, code int, response string) {
+	m.t = t
+	m.code = code
+	m.response = response
+}
+
+// ServeHTTP implements the http.Handler interface.
+func (m *ServerMock) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	require.Equal(m.t, m.path, r.URL.Path)
+	w.WriteHeader(m.code)
+	_, err := io.WriteString(w, m.response)
+	require.NoError(m.t, err)
+}
+
+// NewServerMock builds and returns a
+func NewServerMock(path string) *ServerMock {
+	mock := ServerMock{path: path}
+	mock.Srv = httptest.NewServer(http.HandlerFunc(mock.ServeHTTP))
+	return &mock
+}
+
 func TestVersionServer(t *testing.T) {
 	// Test setup
 	ctx, cancel := context.WithCancel(context.Background())
@@ -91,11 +124,11 @@ func TestVersionServer(t *testing.T) {
 	forwardNoVersionChannel := "forward/none"
 	forwardPath := "/version-server/"
 
-	upstreamServer := basichttp.NewServerMock(forwardPath + constants.VersionPath)
+	upstreamServer := NewServerMock(forwardPath + constants.VersionPath)
 	upstreamServer.SetResponse(t, http.StatusOK, testVersion)
-	upstreamHighServer := basichttp.NewServerMock(forwardPath + constants.VersionPath)
+	upstreamHighServer := NewServerMock(forwardPath + constants.VersionPath)
 	upstreamHighServer.SetResponse(t, http.StatusOK, testVersionMajorTooHigh)
-	upstreamNoVersionServer := basichttp.NewServerMock(forwardPath + constants.VersionPath)
+	upstreamNoVersionServer := NewServerMock(forwardPath + constants.VersionPath)
 	upstreamNoVersionServer.SetResponse(t, http.StatusOK, constants.NoVersion)
 
 	channels := automaticupgrades.Channels{
