+++ lib/automaticupgrades/version/basichttp_test.go
@@ -20,7 +20,9 @@ package version
 
 import (
 	"context"
+	"io"
 	"net/http"
+	"net/http/httptest"
 	"net/url"
 	"testing"
 
@@ -32,10 +34,43 @@ import (
 	"github.com/gravitational/teleport/lib/automaticupgrades/constants"
 )
 
+// ServerMock is a HTTP server whose response can be controlled from the tests.
+// This is used to mock external dependencies like s3 buckets or a remote HTTP server.
+type ServerMock struct {
+	Srv *httptest.Server
+
+	t        *testing.T
+	code     int
+	response string
+	path     string
+}
+
+// SetResponse sets the ServerMock's response.
+func (m *ServerMock) SetResponse(t *testing.T, code int, response string) {
+	m.t = t
+	m.code = code
+	m.response = response
+}
+
+// ServeHTTP implements the http.Handler interface.
+func (m *ServerMock) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	require.Equal(m.t, m.path, r.URL.Path)
+	w.WriteHeader(m.code)
+	_, err := io.WriteString(w, m.response)
+	require.NoError(m.t, err)
+}
+
+// NewServerMock builds and returns a
+func NewServerMock(path string) *ServerMock {
+	mock := ServerMock{path: path}
+	mock.Srv = httptest.NewServer(http.HandlerFunc(mock.ServeHTTP))
+	return &mock
+}
+
 const basicHTTPTestPath = "/v1/cloud-stable"
 
 func Test_basicHTTPVersionClient_Get(t *testing.T) {
-	mock := basichttp2.NewServerMock(basicHTTPTestPath + "/" + constants.VersionPath)
+	mock := NewServerMock(basicHTTPTestPath + "/" + constants.VersionPath)
 	t.Cleanup(mock.Srv.Close)
 	serverURL, err := url.Parse(mock.Srv.URL)
 	serverURL.Path = basicHTTPTestPath
